Move absurd_extreme_comparisons to its own module
